requirements.txt
numpy
torch
opencv-python
pillow
tqdm
scipy
segmentation_models_pytorch
basicsr